Migrate terrifi_firewall_zone off go-unifi SDK (#157)
Adds local internal/unifi.FirewallZone (no DefaultZone field) and
rewrites internal/provider/firewall_zone_api.go to use the local type
directly. The bespoke firewallZoneCreateRequest/UpdateRequest structs
are gone — our local FirewallZone omits the problematic DefaultZone
field at the type level and uses json:"_id,omitempty" so PUT bodies
include the controller-required _id automatically.

Resource, test, and generate files swap from the SDK type. The
file-level commentary in firewall_zone_api.go is reframed: the wire-
format quirks (default_zone serialization, _id in PUT body, 204 on
DELETE, v2-only GET) are controller behavior that any client must
handle, not SDK bugs.
