+++ internal/generate/firewall_zone.go
@@ -1,7 +1,7 @@
 package generate
 
 import (
-	"github.com/ubiquiti-community/go-unifi/unifi"
+	"github.com/alexklibisz/terrifi/internal/unifi"
 )
 
 // FirewallZoneBlocks generates import + resource blocks for firewall zones.
