+++ internal/provider/response_cache_integration_test.go
@@ -12,7 +12,9 @@ import (
 	"github.com/hashicorp/go-retryablehttp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/ubiquiti-community/go-unifi/unifi"
+	sdkunifi "github.com/ubiquiti-community/go-unifi/unifi"
+
+	"github.com/alexklibisz/terrifi/internal/unifi"
 )
 
 // newTestClient creates a Client pointing at the given test server with
@@ -317,9 +319,9 @@ func TestResponseCaching_V1RequestsCachedThroughDoV2(t *testing.T) {
 
 	clientResp := struct {
 		Meta json.RawMessage `json:"meta"`
-		Data []unifi.Client  `json:"data"`
+		Data []sdkunifi.Client  `json:"data"`
 	}{
-		Data: []unifi.Client{
+		Data: []sdkunifi.Client{
 			{ID: "client-1", MAC: "aa:bb:cc:dd:ee:ff", Name: "Test Client"},
 		},
 	}
